Admin middleware imp

+++ src/app/api/menu-items/route.js
@@ -1,17 +1,24 @@
 import { MenuItem } from '@/app/models/MenuItem';
 import mongoose from 'mongoose';
+import { isAdmin } from '../auth/[...nextauth]/route';
 
 export async function POST(req) {
   mongoose.connect(process.env.MONGO_URL_NEXT);
   const data = await req.json();
-  const menuItemDoc = await MenuItem.create(data);
-  return Response.json(menuItemDoc);
+  if (await isAdmin()) {
+    const menuItemDoc = await MenuItem.create(data);
+    return Response.json(menuItemDoc);
+  } else {
+    Response.json({});
+  }
 }
 
 export async function PUT(req) {
   mongoose.connect(process.env.MONGO_URL_NEXT);
-  const { _id, ...data } = await req.json();
-  await MenuItem.findByIdAndUpdate(_id, data);
+  if (await isAdmin()) {
+    const { _id, ...data } = await req.json();
+    await MenuItem.findByIdAndUpdate(_id, data);
+  }
   return Response.json(true);
 }
 
@@ -24,6 +31,8 @@ export async function DELETE(req) {
   mongoose.connect(process.env.MONGO_URL_NEXT);
   const url = new URL(req.url);
   const _id = url.searchParams.get('_id');
-  await MenuItem.deleteOne({ _id });
+  if (await isAdmin()) {
+    await MenuItem.deleteOne({ _id });
+  }
   return Response.json(true);
 }
